feat: advent of code solvers
Some issues with using the npm packages so I moved the output from
wasm-pack into this repo so I can import them relatively

Some issues with how vite handles webworkers.
The relevant pieces of code are commented, including the config change
to the vite internal to astro

Some issues with an error being thrown when navigating from a solver to
any other route, caused by view transitions.
To prevent that I check for a variable in the solver page so the
webworker doesn't get called with an invalid